backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.1.1
httpx==0.27.0
pydantic==2.5.3
aiohttp==3.9.3
playwright==1.41.2
